Replace Secret Service with new Role ID
